refactor kodu + dadanie możliwości sprawdzania znaków podanych przez użytkownika
