+++ .idea/Main.java
@@ -14,6 +14,7 @@ public class Main {
 
         while (kontynuuj){
             if(kontynuuj == false){
+                System.out.println(kontynuuj);
                 break;
             }
             System.out.println("Menu:");
@@ -21,7 +22,6 @@ public class Main {
             System.out.println("2. Wprowadź tekst z pliku");
             System.out.println("3. Pobierz zawartość URL");
             System.out.println("4. Wyjdź");
-            System.out.flush();
 
             int wybor = scanner.nextInt();
             scanner.nextLine(); // Konsumuj znak nowej linii
@@ -49,7 +49,9 @@ public class Main {
                     } catch (IOException e) {
                         System.out.println("Błąd podczas czytania pliku: " + e.getMessage());
                     }
+
                     break;
+
                 case 3:
                     System.out.print("Podaj adres URL: ");
                     String url = scanner.nextLine();
@@ -76,123 +78,28 @@ public class Main {
                     System.out.println("Nieprawidłowy wybór. Wybierz 1, 2, 3 lub 4.");
             }
         }
-        System.out.flush();
+
 
         //Sprawdzanie powtarzalności i zapisywanie o jaki znak chodzi do listy listaZnakow oraz ile razy występuje w liście listaWystapien
-        ArrayList<Character> listaZnakow = new ArrayList<>();
-        ArrayList<Integer> listaWystepowania = new ArrayList<>();
+        List<Character> listaZnakow = new ArrayList<>();
+        List<Integer> listaWystepowania = new ArrayList<>();
 
         for (char litera = 'a'; litera <= 'z'; litera++) {
-            String tekstLower = tekst.toLowerCase();
             int counter = 0;
-            for (char znak : tekstLower.toCharArray()) {
-                if (znak == litera) { counter++; }
+            for (char znak : tekst.toLowerCase().toCharArray()) {
+                if (znak == litera) {
+                    counter++;
+                }
             }
-            if(counter != 0){
-                listaZnakow.add(litera);
-                listaWystepowania.add(counter);
-            }
+            listaZnakow.add(litera);
+            listaWystepowania.add(counter);
         }
-        //Wypisanie ile razy co występuje
+
+        // Wypisz ilość wystąpień
         for (int i = 0; i < listaZnakow.size(); i++) {
-            System.out.println(listaZnakow.get(i) + ": " + listaWystepowania.get(i));
-        }
-
-        kontynuuj = true;
-        while (kontynuuj){
-            if(kontynuuj == false){
-                break;
-            }
-            System.out.println("Menu:");
-            System.out.println("1. Drukuj histogram do konsoli");
-            System.out.println("2. Drukuj histogram do pliku");
-            System.out.println("3. Wyjdź");
-            System.out.flush();
-
-            int wybor = scanner.nextInt(); // pobiera wybór użytkownika
-
-            switch (wybor) {
-                case 1:
-                    GenerujHistogram(listaZnakow, listaWystepowania);
-                    kontynuuj = false;
-                    break;
-                case 2:
-                    System.out.print("Podaj nazwę pliku tekstowego: ");
-                    String nazwaPliku = scanner.nextLine();
-
-                    GenerujHistogram(listaZnakow, listaWystepowania, nazwaPliku);
-                    /*
-                    try {
-                        BufferedReader reader = new BufferedReader(new FileReader(nazwaPliku));
-                        String liniaPliku;
-                        if ((liniaPliku = reader.readLine()) != null) {
-
-                        }
-                        kontynuuj = false;
-                    } catch (FileNotFoundException e) {
-                        System.out.println("Plik nie istnieje: " + e.getMessage());
-                    } catch (IOException e) {
-                        System.out.println("Błąd podczas czytania pliku: " + e.getMessage());
-                    }
-                     */
-                    kontynuuj = false;
-                    break;
-                case 3:
-                    System.out.println("Koniec programu.");
-                    kontynuuj = false;
-                    break;
-                default:
-                    System.out.println("Nieprawidłowy wybór. Wybierz 1, 2 lub 3.");
+            if (listaWystepowania.get(i) != 0) {
+                System.out.println(listaZnakow.get(i) + ": " + listaWystepowania.get(i));
             }
         }
-    }
-    private static void GenerujHistogram(ArrayList<Character> listaZnakow, ArrayList<Integer> listaWystepowania) { // Generowanie histogramu w konsoli
-        System.out.println("Histogram częstości występowania liter:");
-
-        for (int i = 0; i < listaZnakow.size(); i++) {
-            char litera = listaZnakow.get(i);
-            int ilosc = listaWystepowania.get(i);
-            System.out.println(litera + ": " + generujLinieHistogramu(ilosc, ObliczSkale(listaWystepowania)));
-        }
-    }
-    private static void GenerujHistogram(ArrayList<Character> listaZnakow, ArrayList<Integer> listaWystepowania, String nazwaPliku) { // Generowanie histogramu do pliku txt
-        try (PrintWriter writer = new PrintWriter(nazwaPliku)) {
-            writer.println("Histogram częstości występowania liter:");
-
-            for (int i = 0; i < listaZnakow.size(); i++) {
-                char litera = listaZnakow.get(i);
-                int ilosc = listaWystepowania.get(i);
-                writer.println(litera + ": " + generujLinieHistogramu(ilosc, ObliczSkale(listaWystepowania)));
-            }
-
-            System.out.println("Histogram został zapisany do pliku " + nazwaPliku);
-        } catch (FileNotFoundException e) {
-            System.out.println(e);
-        }
-    }
-    private static String generujLinieHistogramu(int ilosc, int skala) {
-        StringBuilder histogram = new StringBuilder();
-        for (int i = 0; i < (ilosc / skala); i++) {
-            histogram.append("*");
-        }
-        return histogram.toString();
-    }
-    private static int ObliczSkale(ArrayList<Integer> listaWystepowania){ // Oblicza skale w jakiej się będzie wyświetlał histogram w przypadku otrzymania dużej liczby danych
-        int skala = 1;
-        if (Max(listaWystepowania) > 20) {
-            skala = Max(listaWystepowania) / 20;
-        }
-        return skala;
-    }
-    public static int Max(ArrayList<Integer> listaWystepowania) { //Zwraca największy element w Array
-        int najwiekszy = listaWystepowania.get(0);
-
-        for (int i = 0; i < listaWystepowania.size(); i++) {
-            if (listaWystepowania.get(i) > najwiekszy) {
-                najwiekszy = listaWystepowania.get(i);
-            }
-        }
-
-        return najwiekszy; // Zwracamy znaleziony największy element.
     }
 }
