+++ .idea/workspace.xml
@@ -1,0 +1,148 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<project version="4">
+  <component name="AutoImportSettings">
+    <option name="autoReloadType" value="SELECTIVE" />
+  </component>
+  <component name="ChangeListManager">
+    <list default="true" id="31359ffa-1252-4dee-9f3e-cbe0c32c4f9f" name="Changes" comment="Działający generator histogramu z dynamiczną skalą. Możliwość wyboru źródła tekstu oraz miejscem generowania histogramu">
+      <change beforePath="$PROJECT_DIR$/.idea/.gitignore" beforeDir="false" afterPath="$PROJECT_DIR$/.gitignore" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/Main.java" beforeDir="false" afterPath="$PROJECT_DIR$/src/Main.java" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/changelog" beforeDir="false" afterPath="$PROJECT_DIR$/changelog" afterDir="false" />
+      <change beforePath="$PROJECT_DIR$/.idea/gitProjekt.iml" beforeDir="false" afterPath="$PROJECT_DIR$/gitProjekt.iml" afterDir="false" />
+    </list>
+    <option name="SHOW_DIALOG" value="false" />
+    <option name="HIGHLIGHT_CONFLICTS" value="true" />
+    <option name="HIGHLIGHT_NON_ACTIVE_CHANGELIST" value="false" />
+    <option name="LAST_RESOLUTION" value="IGNORE" />
+  </component>
+  <component name="Git.Settings">
+    <option name="RECENT_BRANCH_BY_REPOSITORY">
+      <map>
+        <entry key="$PROJECT_DIR$" value="feature1" />
+      </map>
+    </option>
+    <option name="RECENT_GIT_ROOT_PATH" value="$PROJECT_DIR$" />
+  </component>
+  <component name="GitHubPullRequestSearchHistory">{
+  &quot;lastFilter&quot;: {
+    &quot;state&quot;: &quot;OPEN&quot;,
+    &quot;assignee&quot;: &quot;paczuu&quot;
+  }
+}</component>
+  <component name="GithubPullRequestsUISettings">
+    <option name="selectedUrlAndAccountId">
+      <UrlAndAccount>
+        <option name="accountId" value="d8b6c54c-5255-4861-9e57-3ff41b4a5a79" />
+        <option name="url" value="https://github.com/MichalNeuman/gitProjekt.git" />
+      </UrlAndAccount>
+    </option>
+  </component>
+  <component name="MarkdownSettingsMigration">
+    <option name="stateVersion" value="1" />
+  </component>
+  <component name="ProjectColorInfo">{
+  &quot;associatedIndex&quot;: 2
+}</component>
+  <component name="ProjectId" id="2X7NyDYY4g6eAywaWIfUxniGiMD" />
+  <component name="ProjectLevelVcsManager" settingsEditedManually="true" />
+  <component name="ProjectViewState">
+    <option name="hideEmptyMiddlePackages" value="true" />
+    <option name="showLibraryContents" value="true" />
+  </component>
+  <component name="PropertiesComponent"><![CDATA[{
+  "keyToString": {
+    "RunOnceActivity.OpenProjectViewOnStart": "true",
+    "RunOnceActivity.ShowReadmeOnStart": "true",
+    "SHARE_PROJECT_CONFIGURATION_FILES": "true",
+    "WebServerToolWindowFactoryState": "false",
+    "git-widget-placeholder": "feature2",
+    "node.js.detected.package.eslint": "true",
+    "node.js.detected.package.tslint": "true",
+    "node.js.selected.package.eslint": "(autodetect)",
+    "node.js.selected.package.tslint": "(autodetect)",
+    "nodejs_package_manager_path": "npm",
+    "vue.rearranger.settings.migration": "true"
+  }
+}]]></component>
+  <component name="RecentsManager">
+    <key name="MoveFile.RECENT_KEYS">
+      <recent name="C:\Users\patry\IdeaProjects\gitProjekt" />
+      <recent name="C:\Users\patry\IdeaProjects\gitProjekt\src" />
+    </key>
+  </component>
+  <component name="SpellCheckerSettings" RuntimeDictionaries="0" Folders="0" CustomDictionaries="0" DefaultDictionary="application-level" UseSingleDictionary="true" transferred="true" />
+  <component name="TaskManager">
+    <task active="true" id="Default" summary="Default task">
+      <changelist id="31359ffa-1252-4dee-9f3e-cbe0c32c4f9f" name="Changes" comment="" />
+      <created>1697976410684</created>
+      <option name="number" value="Default" />
+      <option name="presentableId" value="Default" />
+      <updated>1697976410684</updated>
+      <workItem from="1697976411785" duration="1137000" />
+      <workItem from="1697977573165" duration="1642000" />
+      <workItem from="1697980396630" duration="9000" />
+      <workItem from="1697980423646" duration="8000" />
+      <workItem from="1697980518214" duration="350000" />
+      <workItem from="1697980884316" duration="509000" />
+      <workItem from="1697982552735" duration="203000" />
+      <workItem from="1697982768112" duration="4000" />
+      <workItem from="1697982783886" duration="1633000" />
+      <workItem from="1697985553404" duration="5000" />
+      <workItem from="1697985567111" duration="280000" />
+    </task>
+    <task id="LOCAL-00001" summary="Dodano generacje histogramu + tworzenie switcha aby wybrać czy histogram ma się drukować w konsoli czy do pliku">
+      <option name="closed" value="true" />
+      <created>1697977626543</created>
+      <option name="number" value="00001" />
+      <option name="presentableId" value="LOCAL-00001" />
+      <option name="project" value="LOCAL" />
+      <updated>1697977626543</updated>
+    </task>
+    <task id="LOCAL-00002" summary="Dodano generacje histogramu + tworzenie switcha aby wybrać czy histogram ma się drukować w konsoli czy do pliku v2">
+      <option name="closed" value="true" />
+      <created>1697981367314</created>
+      <option name="number" value="00002" />
+      <option name="presentableId" value="LOCAL-00002" />
+      <option name="project" value="LOCAL" />
+      <updated>1697981367314</updated>
+    </task>
+    <task id="LOCAL-00003" summary="Działający generator histogramu z dynamiczną skalą. Możliwość wyboru źródła tekstu oraz miejscem generowania histogramu">
+      <option name="closed" value="true" />
+      <created>1697982729910</created>
+      <option name="number" value="00003" />
+      <option name="presentableId" value="LOCAL-00003" />
+      <option name="project" value="LOCAL" />
+      <updated>1697982729910</updated>
+    </task>
+    <task id="LOCAL-00004" summary="Działający generator histogramu z dynamiczną skalą. Możliwość wyboru źródła tekstu oraz miejscem generowania histogramu">
+      <option name="closed" value="true" />
+      <created>1697982833824</created>
+      <option name="number" value="00004" />
+      <option name="presentableId" value="LOCAL-00004" />
+      <option name="project" value="LOCAL" />
+      <updated>1697982833825</updated>
+    </task>
+    <option name="localTasksCounter" value="5" />
+    <servers />
+  </component>
+  <component name="TypeScriptGeneratedFilesManager">
+    <option name="version" value="3" />
+  </component>
+  <component name="Vcs.Log.Tabs.Properties">
+    <option name="TAB_STATES">
+      <map>
+        <entry key="MAIN">
+          <value>
+            <State />
+          </value>
+        </entry>
+      </map>
+    </option>
+  </component>
+  <component name="VcsManagerConfiguration">
+    <MESSAGE value="Dodano generacje histogramu + tworzenie switcha aby wybrać czy histogram ma się drukować w konsoli czy do pliku" />
+    <MESSAGE value="Dodano generacje histogramu + tworzenie switcha aby wybrać czy histogram ma się drukować w konsoli czy do pliku v2" />
+    <MESSAGE value="Działający generator histogramu z dynamiczną skalą. Możliwość wyboru źródła tekstu oraz miejscem generowania histogramu" />
+    <option name="LAST_COMMIT_MESSAGE" value="Działający generator histogramu z dynamiczną skalą. Możliwość wyboru źródła tekstu oraz miejscem generowania histogramu" />
+  </component>
+</project>
